Fix host-check: LP64 types, circular include, missing DC stubs
- BPEEnvironment.h: typedef uint32 as unsigned int on LP64 (unsigned long
  is 8 bytes on x86_64, breaking TRenderHandle size assertion); add <stdint.h>
  for uintptr_t
- BPEAssert.cpp: match implementation signature (uint32) to header declaration
  (was unsigned long, causing overload ambiguity after LP64 fix)
- DCCRenderBackend.h: remove circular #include of CRenderBackend.h; platform
  plugin headers must not include the base header that includes them
- DCCRenderBackend.cpp: include CRenderBackend.h instead of DCCRenderBackend.h
  directly (matches VTA pattern); initialize CMatrix4 members in ctor
- DCCCompiledShader.h: add Begin/End/BeginPass/EndPass/BeginShaderBatch/
  EndShaderBatch stubs (PSCShader.cpp calls these on all ProgShader targets)
- DCCVertexData.h: create (mirrors VTACVertexData; required by CVertexData.h
  platform include macro)
- CMaterialFlags.h: create stub (used throughout ProgShader shaders but never
  defined in this codebase)
- CCPLdrShaderUnlitGeneric.h: create stub (required by CUnlitGeneric.cpp)
- PSCShader.h: include CMaterialFlags.h so all ProgShader shaders see it
- CDrawableRenderEntity.h: forward-declare CMatrix34 (used in virtual methods)
- CMeshRenderEntity.h: include CMatrix4.h (CMatrix34.h only forward-declares it)
- CShaderParameterBuffer.h/cpp: add 2-arg AddTexture overload; fix pointer-to-
  uint32 cast via uintptr_t for LP64 host compatibility
- CVertexArray.cpp: fix void*/uint8* to uint32 casts via uintptr_t
- CResourceManager.cpp: fix uint32*/uLongf* mismatch in uncompress_gzip via
  local uLongf variable
- Makefile.dc: drop CUnlitGeneric.cpp, CVertexLitGeneric.cpp, PSCSkinnedMesh.cpp
  (depend on game-specific types absent from this codebase)

+++ bp/Source/Engine/Basics/BPEAssert.cpp
@@ -312,7 +312,7 @@ void bpe_add_string_part( char * msgString, int const maxMsgStringSize, int & ms
 
 //----------------------------------------------------------------------------------------------
 
-void bpe_log_assert_failure(char const * const fileName, unsigned long lineNum,
+void bpe_log_assert_failure(char const * const fileName, uint32 lineNum,
                        char const * const prefix, char const * const cond, char const * const msg )
 {
    printf(
@@ -329,7 +329,7 @@ void bpe_log_assert_failure(char const * const fileName, unsigned long lineNum,
 
 //----------------------------------------------------------------------------------------------
 
-bool bpe_query_to_halt_program(char const * const fileName, unsigned long const lineNum,
+bool bpe_query_to_halt_program(char const * const fileName, uint32 const lineNum,
                           char const * const prefix, char const * const cond, char const * const msg)
 {
